change score type from float to int

+++ exercises/exercise1/src/FullTimeStudent.java
@@ -1,5 +1,5 @@
 public class FullTimeStudent extends Student {
-    private float examScore[] = new float[2];
+    private Integer examScore[] = new Integer[2];
 
     public FullTimeStudent(String name) {
         super(name);
@@ -12,11 +12,11 @@ public class FullTimeStudent extends Student {
     public float getExamScore2() {
         return this.examScore[1];
     }
-    public void setExamScore1(float score1) {
+    public void setExamScore1(int score1) {
         examScore[0] = score1;
     }
 
-    public void setExamScore2(float score2) {
+    public void setExamScore2(int score2) {
         examScore[1] = score2;
     }
 }
